refactor(web): move web3 utils to @safe-global/utils
