+++ apps/web/src/services/security/modules/ApprovalModule/index.ts
@@ -4,7 +4,7 @@ import {
   INCREASE_ALLOWANCE_SIGNATURE_HASH,
 } from '@/components/tx/ApprovalEditor/utils/approvals'
 import { ERC20__factory } from '@/types/contracts'
-import { normalizeTypedData } from '@/utils/web3'
+import { normalizeTypedData } from '@safe-global/utils/utils/web3'
 import { type SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import { id } from 'ethers'
 import { type SecurityResponse, type SecurityModule, SecuritySeverity } from '../types'
