+++ apps/web/src/services/security/modules/BlockaidModule/index.ts
@@ -1,6 +1,6 @@
 import type { TypedData } from '@safe-global/store/gateway/AUTO_GENERATED/messages'
 import { isEIP712TypedData } from '@/utils/safe-messages'
-import { normalizeTypedData } from '@/utils/web3'
+import { normalizeTypedData } from '@safe-global/utils/utils/web3'
 import { type SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import { generateTypedData } from '@safe-global/protocol-kit/dist/src/utils/eip-712'
 import { type SecurityResponse, type SecurityModule, SecuritySeverity } from '../types'
